requirements.txt
flask==3.0.3
flask_sqlalchemy==3.1.1
sqlalchemy==2.0.32
flask-login==0.6.3
scikit-learn==1.5.1
pandas==2.2.2
numpy==1.26.4
python-docx==1.1.2
pdfminer.six==20240706
